pyproject.toml
[project]
name = "gen-gec-errant"
version = "0.1.0"
description = "Pipeline for automatic text generation, grammatical error correction, and ERRANT error annotation"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "errant>=3.0.0",
    "spacy>=3.5.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
llm-gec = ["accelerate>=0.25.0"]

[project.scripts]
gen-gec-errant = "gen_gec_errant.pipeline.__main__:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
